requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
pydantic
PyJWT
pydantic-settings